hw3/requirements.txt
opencv-python==4.5.3.56
numpy
scipy
matplotlib
Pillow
